refactor(apps): move browser.ts to app root, symmetric to daemon.ts
`browser.ts` lived four levels deep under SvelteKit's signed-in route
because it grew out of that route. Its job is browser-side runtime
composition around `openFooWorkspace`, which is the browser counterpart
to `daemon.ts`. Put it next to `daemon.ts` and `workspace.ts` at the
app root so the three-file boundary is visible at a glance:

  apps/fuji/workspace.ts   shared opener and schema
  apps/fuji/browser.ts     browser runtime composition
  apps/fuji/daemon.ts      daemon extension composition

Route files keep their UI responsibilities; importers update to point at
`../../browser`. Browser.ts now consumes `workspace.actions` (folded in
the previous commit) instead of calling the factory itself, and gains a
comment naming the intentional per-rich-text-body WebSocket: each body
is its own Y.Doc / sync room, so a per-body open is by design.
